+++ _example/beego.go
