+++ _example/echo.go
@@ -3,7 +3,7 @@ package example
 import (
 	"net/http"
 
-	"github.com/labstack/echo"
+	"github.com/labstack/echo/v4"
 )
 
 func echoHelloHandler() echo.HandlerFunc {
