+++ _example/pat.go
