+++ _example/pat_test.go
