+++ _example/tango.go
