+++ go.mod
@@ -1,21 +1,3 @@
 module github.com/appleboy/gofight/v2
 
-go 1.12
-
-require (
-	gitea.com/lunny/tango v0.6.1
-	github.com/astaxie/beego v1.11.1
-	github.com/gin-gonic/gin v1.4.0
-	github.com/gorilla/context v1.1.1 // indirect
-	github.com/gorilla/mux v1.7.2
-	github.com/gorilla/pat v0.0.0-20180118222023-199c85a7f6d1
-	github.com/julienschmidt/httprouter v1.2.0
-	github.com/labstack/echo v3.3.10+incompatible
-	github.com/labstack/gommon v0.2.8 // indirect
-	github.com/mattn/go-colorable v0.1.2 // indirect
-	github.com/stretchr/testify v1.3.0
-	github.com/tidwall/gjson v1.2.1
-	github.com/tidwall/match v1.0.1 // indirect
-	github.com/tidwall/pretty v0.0.0-20190325153808-1166b9ac2b65 // indirect
-	github.com/valyala/fasttemplate v1.0.1 // indirect
-)
+go 1.13
